pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "third-chair"
version = "0.1.0"
description = "Legal discovery processing tool for Axon evidence packages"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "David Karpay" }
]
keywords = ["legal", "discovery", "transcription", "translation", "axon"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",

    # Transcription
    "faster-whisper>=1.0.0",
    "soundfile>=0.12.1",
    "ffmpeg-python>=0.2.0",

    # Diarization (optional - requires HF_TOKEN)
    "pyannote.audio>=3.0.0",
    "torch>=2.0.0",

    # Translation / Language Detection
    "transformers>=4.40.0",
    "sentencepiece>=0.1.99",
    "fasttext-wheel>=0.9.2",
    "httpx>=0.27.0",  # For Ollama API

    # Document Processing
    "pdfplumber>=0.10.0",
    "pytesseract>=0.3.10",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",

    # Report Generation
    "reportlab>=4.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
third-chair = "third_chair.cli.main:app"

[project.urls]
Repository = "https://github.com/davidkarpay/Third-Chair"

[tool.hatch.build.targets.wheel]
packages = ["third_chair"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
